refactor: Extract the valueType checking func as validation Tag
Fix #383
